+++ docs/changelog.md
@@ -213,3 +213,4 @@
 - Split the unpause functionality into a seperate method so that other methods can use it if need be.
 - Added variable M_GameOver to track if the game is over or not.
 - Added M_GameOver to be false as a condition of the pause menu appearing.
+- Fixed a bug with the Fruit timeline where it wouldnt stop after the bonus fruit was eaten.